refactor: move redis implementation out of data store module since we will use it for tokens in the future too
